refactor: consolidate state management into app/stores/
- Move app/state/AppState.js to app/stores/AppState.js
- Move app/state/README.md to app/stores/README.md
- Remove app/state/ directory
- Update all imports and documentation references
- Verify with full test suite
